Left sidebar, change password and small fixes

+++ admin/advertisement_delete.php
@@ -23,7 +23,7 @@ if(isset($_GET['id']) && $_GET['id']!="")
 	$row	= mysql_fetch_assoc($sql);
 	$image_loc	= $row['image_name'];
 	
-	if($image_loc != "" )	
+	if($image_loc != "" && file_exists($image_loc))	
 		$img_del_result = unlink($image_loc);
 	$sql=mysql_query("DELETE FROM cashbackengine_advertisements WHERE advertisement_id='$id'");
